refactor(cdn): move s3 config to shared
