+++ shared/src/config.rs
@@ -102,3 +102,41 @@ pub struct EventStreamConfig {
 	#[settings(default = "seventv".into())]
 	pub stream_prefix: String,
 }
+
+#[auto_settings]
+#[serde(default)]
+pub struct S3BucketConfig {
+	/// The name of the S3 bucket
+	#[settings(default = String::from("7tv-public"))]
+	pub name: String,
+	/// The region the S3 bucket is in
+	#[settings(default = String::from("us-east-1"))]
+	pub region: String,
+	/// The custom endpoint for the S3 bucket
+	#[settings(default = Some("http://localhost:9000".to_string()))]
+	pub endpoint: Option<String>,
+	/// The credentials for the S3 bucket
+	#[settings(default = S3CredentialsConfig::default())]
+	pub credentials: S3CredentialsConfig,
+}
+
+#[auto_settings]
+#[serde(default)]
+pub struct S3CredentialsConfig {
+	/// The access key for the S3 bucket
+	pub access_key: Option<String>,
+	/// The secret key for the S3 bucket
+	pub secret_key: Option<String>,
+}
+
+impl S3CredentialsConfig {
+	pub fn to_credentials(&self) -> Option<aws_sdk_s3::config::Credentials> {
+		Some(aws_sdk_s3::config::Credentials::new(
+			self.access_key.clone()?,
+			self.secret_key.clone()?,
+			None,
+			None,
+			"seventv",
+		))
+	}
+}
